Add deflate method recognition
This commit modifies the inflate source code to report that it
recognizes compression method 8 for DEFLATE. In addition, the inflate
source code outputs dummy data, for testing.

The z reader actually sends a put callback for the inflate source to
use. In addition, the z reader now properly handles end of compressed
DEFLATE stream from the compression callbacks.